# docs/swagger.yaml
basePath: /reporting-api
definitions:
  models.AuthUserData:
    properties:
      authStatus:
        type: integer
      displayName:
        type: string
      domain:
        type: string
      name:
        type: string
      permission:
        items:
          $ref: '#/definitions/models.MyPermission'
        type: array
      type:
        type: string
    type: object
  models.CreateReport:
    properties:
      alias:
        type: string
      description:
        type: string
      operation_name:
        type: string
      owner:
        type: string
      parentId:
        type: integer
      privateAlias:
        type: string
      privateAliasExpiresAt:
        type: string
      text:
        type: string
      url:
        type: string
      visible:
        type: boolean
    type: object
  models.DataField:
    properties:
      url:
        type: string
    type: object
  models.FavoriteReport:
    properties:
      categoryId:
        type: integer
      id:
        type: integer
      login:
        type: string
      reportId:
        type: integer
    type: object
  models.FavoriteReportDTO:
    properties:
      categoryId:
        type: integer
      reportId:
        type: integer
    required:
    - categoryId
    - reportId
    type: object
  models.GetCategory:
    properties:
      alias:
        type: string
      description:
        type: string
      id:
        type: integer
      parentId:
        type: integer
      privateAlias:
        type: string
      privateAliasExpiresAt:
        type: string
      text:
        type: string
      type:
        type: string
      visible:
        type: boolean
    type: object
  models.InsertCategory:
    properties:
      alias:
        type: string
      description:
        type: string
      parentId:
        type: integer
      privateAlias:
        type: string
      privateAliasExpiresAt:
        type: string
      text:
        type: string
      visible:
        type: boolean
    type: object
  models.MyPermission:
    properties:
      name:
        type: string
      permission:
        type: integer
    type: object
  models.PathDataResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.GetCategory'
        type: array
      isPrivate:
        type: boolean
      reports:
        items:
          $ref: '#/definitions/models.Report'
        type: array
    type: object
  models.Report:
    properties:
      alias:
        type: string
      data:
        $ref: '#/definitions/models.DataField'
      description:
        type: string
      id:
        type: integer
      operation_name:
        type: string
      owner:
        type: string
      parentId:
        type: integer
      privateAlias:
        type: string
      privateAliasExpiresAt:
        type: string
      text:
        type: string
      type:
        type: string
      visible:
        type: boolean
    type: object
  models.TrackVisitDTO:
    properties:
      categoryId:
        type: integer
      reportId:
        type: integer
    required:
    - categoryId
    - reportId
    type: object
  models.UpdateCategory:
    properties:
      description:
        type: string
      id:
        type: integer
      privateAlias:
        type: string
      privateAliasExpiresAt:
        type: string
      text:
        type: string
      visible:
        type: boolean
    type: object
  models.UpdateCategoryParent:
    properties:
      fromCat:
        type: integer
      id:
        type: integer
      toCat:
        type: integer
    type: object
  models.UpdateReport:
    properties:
      description:
        type: string
      id:
        type: integer
      operation_name:
        type: string
      owner:
        type: string
      privateAlias:
        type: string
      privateAliasExpiresAt:
        type: string
      text:
        type: string
      url:
        type: string
      visible:
        type: boolean
    type: object
  models.UserData:
    properties:
      domain:
        type: string
      login:
        type: string
      password:
        type: string
    type: object
  models.VisitedReport:
    properties:
      alias:
        type: string
      category_alias:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server for reporting API.
  title: Reporting API
  version: "1.0"
paths:
  /api/Authorization/GetCurrentUserInfo:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthUserData'
      tags:
      - Authorization
  /api/Authorization/LogInAuthorization:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: userdata
        required: true
        schema:
          $ref: '#/definitions/models.UserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthUserData'
      tags:
      - Authorization
  /api/Authorization/LogOutAuthorization:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Authorization
  /api/Categories:
    get:
      consumes:
      - application/json
      description: Получение списка категорий
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetCategory'
            type: array
      summary: Get Categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Создать категорию
      parameters:
      - description: Новая категория
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.InsertCategory'
      produces:
      - application/json
      responses:
        "200":
          description: Created Category Id
          schema:
            type: integer
      summary: Insert Category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Обновить категорию
      parameters:
      - description: Обновленная категория
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update Category
      tags:
      - Categories
  /api/Categories/{categoryId}:
    delete:
      consumes:
      - application/json
      description: Удалить категорию
      parameters:
      - description: Id категории
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete Category
      tags:
      - Categories
  /api/Categories/GetCategoriesForAdmin:
    get:
      consumes:
      - application/json
      description: Получение списка категорий для администратора
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetCategory'
            type: array
      summary: Get Categories for Admin
      tags:
      - Categories
  /api/Categories/UpdateCategoryParent:
    put:
      consumes:
      - application/json
      description: Обновить родительскую категорию
      parameters:
      - description: Обновленный родительский айди
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryParent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update Parent Category
      tags:
      - Categories
  /api/FavoriteReports/AddReport:
    post:
      consumes:
      - application/json
      description: Add a report to favorites by report ID and category ID
      parameters:
      - description: Favorite report request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.FavoriteReportDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a report to favorites
      tags:
      - FavoriteReports
  /api/FavoriteReports/DeleteReport:
    delete:
      consumes:
      - application/json
      description: Remove a report from favorites by report ID and category ID
      parameters:
      - description: Favorite report request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.FavoriteReportDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a report from favorites
      tags:
      - FavoriteReports
  /api/FavoriteReports/GetReports:
    get:
      description: Get a list of all favorite reports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FavoriteReport'
            type: array
      summary: Get all favorite reports
      tags:
      - FavoriteReports
  /api/Reports:
    get:
      description: Get a list of all reports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Report'
            type: array
      summary: Get all reports
      tags:
      - Reports
    post:
      consumes:
      - application/json
      description: Create a new report
      parameters:
      - description: Report to create
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/models.CreateReport'
      produces:
      - application/json
      responses:
        "200":
          description: Created report ID
          schema:
            type: integer
      summary: Create a new report
      tags:
      - Reports
    put:
      consumes:
      - application/json
      description: Update an existing report
      parameters:
      - description: Report to update
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/models.UpdateReport'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Update a report
      tags:
      - Reports
  /api/Reports/{reportId}:
    delete:
      consumes:
      - application/json
      description: Remove a report from a category
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: integer
      - description: Category ID
        in: body
        name: categoryId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Remove a report
      tags:
      - Reports
  /api/Reports/AddReportRelation/{reportId}:
    post:
      consumes:
      - application/json
      description: Add a report to a category
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: integer
      - description: Category ID
        in: body
        name: categoryId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Add report relation
      tags:
      - Reports
  /api/Reports/GetReportsForAdmin:
    get:
      consumes:
      - application/json
      description: Получение списка отчетов для администратора
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Report'
            type: array
      summary: Get Reports for Admin
      tags:
      - Reports
  /api/Reports/UpdateCategoryReports:
    put:
      consumes:
      - application/json
      description: Update the category of a report
      parameters:
      - description: Category data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryParent'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Update category reports
      tags:
      - Reports
  /api/TrackVisit:
    post:
      consumes:
      - application/json
      description: Track visited reports
      parameters:
      - description: Report and Category IDs
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.TrackVisitDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add visited report
      tags:
      - VisitHistory
  /api/TrackVisit/GetLastVisited:
    get:
      description: Get last visited reports
      parameters:
      - default: 20
        description: Quantity of reports to return
        in: query
        name: quantity
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.VisitedReport'
            type: array
      summary: Get last visited reports
      tags:
      - VisitedReports
  /api/data:
    get:
      consumes:
      - application/json
      description: Returns all visible categories and reports (empty path or two aliases)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PathDataResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all visible categories and reports
      tags:
      - data
  /api/data/{alias}:
    get:
      consumes:
      - application/json
      description: Returns categories and reports for a single alias (public or private)
      parameters:
      - description: Alias (category or private alias)
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PathDataResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category or report by single alias
      tags:
      - data
  /api/data/{alias}/{subalias}:
    get:
      consumes:
      - application/json
      description: Returns all visible categories and reports (empty path or two aliases)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PathDataResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all visible categories and reports
      tags:
      - data
swagger: "2.0"